+++ eyeball/benches/set_a_lot.rs
@@ -2,7 +2,7 @@
 
 use criterion::{black_box, criterion_group, criterion_main, BatchSize, Bencher, Criterion};
 
-use eyeball::unique::Observable;
+use eyeball::Observable;
 use tokio::task::JoinSet;
 
 fn baseline(b: &mut Bencher<'_>) {