+++ eyeball/CHANGELOG.md
@@ -1,11 +1,15 @@
 # unreleased
 
-- Add a new generic parameter to `Subscriber` and `shared::Observable` that
+- Rename / move the observable types:
+  - `eyeball::unique::Observable` is now `eyeball::Observable`
+  - `eyeball::shared::Observable` is now `eyeball::SharedObservable`
+- Add a new generic parameter to `SharedObservable` and `Subscriber`¹ that
   controls whether the internal lock is an async-aware one or not. It defaults
   to `SyncLock` which is the same behavior as before, but can be set to
   `AsyncLock` (created with `Observable::new_async`), if you want to lock the
   inner value for writing over `.await` points in async code. This means that
-  most operations on the observable and its subscribers become `async`.
+  most operations on the observable and its subscribers become `async`.\
+  ¹ also for `Observable`, but much less useful there
 
 # 0.7.0